# Corrected Throughput measure for GaudiDDPMPipeline