… tests/test_diffusers.py