# What does this PR do?

The test GaudiDDPMPipelineTester::test_no_throughput_regression_bf16() fails with throughput comparison, but throughput was being measured inversely as sec/sample instead of samples /sec.
This PR does the follow

- Corrected throughput measure formula for GaudiDDPMPipeline
- Updated the Gaudi benchmark values for diffusion test test_no_throughput_regression_bf16()


Below is the summary of slow_diffusion tests

python -m pytest tests/test_diffusers.py -v -s -k "test_textual_inversion"
============================= test session starts ==============================
platform linux -- Python 3.10.12, pytest-7.4.4, pluggy-1.5.0 -- /usr/bin/python
cachedir: .pytest_cache
rootdir: /mnt/oh_pr
configfile: setup.cfg
collecting ... collected 159 items / 158 deselected / 1 selected
========== 1 passed, 158 deselected, 5 warnings in 160.91s (0:02:40) ===========


python -m pytest tests/test_diffusers.py -v -s -k "test_train_text_to_image_"
============================= test session starts ==============================
platform linux -- Python 3.10.12, pytest-7.4.4, pluggy-1.5.0 -- /usr/bin/python
cachedir: .pytest_cache
rootdir: /mnt/oh_pr
configfile: setup.cfg
collecting ... collected 159 items / 157 deselected / 2 selected
========== 2 passed, 157 deselected, 3 warnings in 486.10s (0:08:06) ===========


python -m pytest tests/test_diffusers.py -v -s -k "test_train_controlnet"
============================= test session starts ==============================
platform linux -- Python 3.10.12, pytest-7.4.4, pluggy-1.5.0 -- /usr/bin/python
cachedir: .pytest_cache
rootdir: /mnt/oh_pr
configfile: setup.cfg
collecting ... collected 159 items / 157 deselected / 2 selected
========== 2 passed, 157 deselected, 5 warnings in 201.56s (0:03:21) ===========


python -m pytest tests/test_diffusers.py -v -s -k "test_deterministic_image_generation"
============================= test session starts ==============================
platform linux -- Python 3.10.12, pytest-7.4.4, pluggy-1.5.0 -- /usr/bin/python
cachedir: .pytest_cache
rootdir: /mnt/oh_pr
configfile: setup.cfg
collecting ... collected 159 items / 157 deselected / 2 selected
========== 2 passed, 157 deselected, 6 warnings in 276.77s (0:04:36) ===========


python -m pytest tests/test_diffusers.py -v -s -k "test_no_"
============================= test session starts ==============================
platform linux -- Python 3.10.12, pytest-7.4.4, pluggy-1.5.0 -- /usr/bin/python
cachedir: .pytest_cache
rootdir: /mnt/oh_pr
configfile: setup.cfg
collecting ... collected 159 items / 153 deselected / 6 selected
===== 5 passed, 1 skipped, 153 deselected, 6 warnings in 416.68s (0:06:56) =====


## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [ ] Did you make sure to update the documentation with your changes?
- [x] Did you write any new necessary tests?