@yafshar 2024-10-28
@deepak-gowda-narayana nice catch! You are right throughput typically measures the number of units processed per unit of time and here it is inverse of that! 

Would you please post the tests and failures before this PR?