@deepak-gowda-narayana 2024-10-28
> @deepak-gowda-narayana nice catch! You are right throughput typically measures the number of units processed per unit of time and here it is inverse of that!
> 
> Would you please post the tests and failures before this PR?

@yafshar 

Sure, here is the test summary

Result Summary of fast_tests.sh

python -m pytest tests/test_gaudi_configuration.py tests/test_trainer_distributed.py tests/test_trainer.py tests/test_trainer_seq2seq.py
=============================================================================================== 81 passed, 8 skipped, 39 warnings in 109.40s (0:01:49) ================================================================================================

Result Summary of fast_tests_diffusers.sh

python -m pytest tests/test_diffusers.py
================================================================================================= 113 passed, 47 skipped, 280 warnings in 1229.45s (0:20:29) =================================================================================================

Result Summary of slow_tests_diffusers.sh
================================================================================================================== short test summary info ===================================================================================================================
FAILED tests/test_diffusers.py::GaudiDDPMPipelineTester::test_no_throughput_regression_bf16 - AssertionError: 6.937816172838211 not greater than or equal to 7.287651444971561
======================================================================================= 1 failed, 4 passed, 1 skipped, 154 deselected, 6 warnings in 358.13s (0:05:58) =======================================================================================
make: *** [Makefile:104: slow_tests_diffusers] Error 1